pyproject.toml
[project]
name = "music-emotion-probing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flax>=0.10.6",
    "h5py>=3.14.0",
    "jax>=0.6.2",
    "jaxlib>=0.6.2",
    "matplotlib>=3.10.3",
    "numpy>=2.3.1",
    "optax>=0.2.5",
    "orbax-checkpoint>=0.11.16",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "pytest>=8.4.1",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.7.0",
    "scipy>=1.16.0",
    "seaborn>=0.13.2",
    "timesfm>=1.2.9",
    "torch>=2.7.1",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "transformers>=4.52.4",
]
